+++ src/entities/post/ui/PostTable.tsx
@@ -41,7 +41,6 @@ const PostTable: React.FC<PostTableProps> = ({
   onTagClick,
   renderUserAvatar,
 }) => {
-  console.log('posts', posts);
   return (
     <Table>
       <TableHeader>
