+++ src/widgets/comment-management/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as CommentWidget } from './CommentWidget';
