+++ src/widgets/index.ts
@@ -1,0 +1,3 @@
+export * from './post-manager';
+export * from './comment-management';
+export * from './user-profile';
