+++ src/widgets/post-manager/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as PostManagerWidget } from './PostManagerWidget';
