+++ src/widgets/user-profile/ui/UserProfileWidget.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import { UserModalManager } from '../../../features/user-profile';
+
+const UserProfileWidget: React.FC = () => {
+  return (
+    <>
+      {/* 사용자 프로필 모달 */}
+      <UserModalManager />
+    </>
+  );
+};
+
+export default UserProfileWidget;
